backend/requirements.txt
# Core framework — 5.2 is LTS, supported until April 2028
Django==5.2

# REST API
djangorestframework==3.15.2

# CORS — allow frontend (Nginx) to call backend
django-cors-headers==4.4.0

# PostgreSQL driver — psycopg3 (modern, production-safe, async-capable)
# Replaces psycopg2-binary which is explicitly discouraged in production
psycopg[binary]==3.2.3

# JWT authentication
djangorestframework-simplejwt==5.3.1

# Argon2id password hasher — memory-hard, GPU-resistant (replaces PBKDF2 default)
# See PASSWORD_HASHERS in settings/base.py
argon2-cffi==23.1.0

# Query filtering support for DRF list views
django-filter==24.2

# Environment variable management (replaces unmaintained python-decouple)
django-environ==0.11.2

# WSGI server — latest stable with security fixes
gunicorn==23.0.0
